// src/core/model/staff/newstaff.ts
export interface newStaffRequest {
  firstName: string;
  lastName: string;
  middleName: string;
  address: string;
  email: string;
  occupation: string;
  dob: string;
  gender: string;
  yearsOfExperience: string;
  medicalLicense: string;
  employmentStatus: string;
  telephoneNumber: string;
}